+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -20,6 +20,9 @@ Notes](../../RELEASENOTES.md).
   timeout.
   ([#7447](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7447))
 
+* Fixed the OTLP/HTTP exporter silently dropping data when an export timed out.
+  ([#7455](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7455))
+
 ## 1.16.0
 
 Released 2026-Jun-10